Support 'strange' memory sizes, such as 256K, 192K
Additional ROM patches manually set the memTop variable to a given
memory size, and the memory test routine is skipped.  This allows
the Plus ROM to cope with a Mac 256K configuration, but also non-
power-of-two sizes.  That's useful for an MCU, making a Mac 208K,
Mac 192K, etc. possible.

Memory addresses are clamped such that wrapping within the RAM
region works, and any emulated address can be used without
overflowing the host RAM allocation.  For a power-of-two memory
size this boils down (after compiler optimisation) to a plain
tmask; for non-Po2, this involves a divide/mod in the weird ROM
probing "off the end" cases, and a simple compare in all other
runtime cases (so shouldn't be too slow).