discs: Add write support
The unix_main gets another option, -w, that enables the disc
mapping to reflect writes back to the backing file.

The disc emulation code now handles writes, which are copied
back to the disc data.

This still doesn't abstract read/write ops to disc, which will
be the next step (to support, for example, block-based storage).