Support discontiguous host RAM
Particularly useful for pico-umac, which wants to keep the FB in
internal memory.  By defining RAM_SIZE_HI, a split is enabled
by ram base calculations.

This has a couple of corner cases to consider:
- The boundary should be at least aligned to 4 bytes (and probably
  kilobytes!) as that then prevents a 68k access straddling the
  boundary, which won't work.
- A disc read/write could straddle the boundary; this is dealt
  with in a later commit.

When RAM_SIZE_HI is defined, a new umac_init_ext() init function
is used to supply the base address of the high region.

Otherwise, everthing is backward-compatible.

NOTE: There isn't a build/Makefile configuration for this yet.
There's no real advantage (aside from testing) to use this config
with the regular SDL/unix_main build; it's intended for pico-umac.