disc: Refactor read/write routines, support discontiguous accesses
When discontiguous memory support is built (RAM_SIZE_HI, etc.),
a disc R/W could straddle the boundary between two regions of memory.

This has a number of cases, e.g. a set of blocks before the split,
straddling the split, and after the split.

This code seems to work (reads have been tested in all three cases,
writes only the straddling case), but should be refactored to deal
with all reads/writes in the same function.  (See FIXME.)